pdf_service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pypdf2